requirements.txt
requests
sqlalchemy
mysql-connector-python
pyspark